move proto/pingping to discovery module
This is needed for the new peer discovery enabled by libp2p
